fix: challengAmount/Count 기본값 0으로 설정( #199)

+++ src/main/java/com/savit/challenge/dto/ParticipantInfo.java
@@ -18,7 +18,7 @@ public class ParticipantInfo {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     private BigDecimal challengeAmount;
     @JsonInclude(JsonInclude.Include.NON_NULL)
-    private int challengeCount;
+    private Long challengeCount;
 
     private String status;
